Fix expired notice for recent expiry dates
If an expiry date is in the recent past then the difference in months will still return zero, and it will be shown as soon to expire.

Replace the check of months difference `< 0` with a call to `isPast` to ensure all expiry dates in the past show correctly as expired.