Support AutoComplete usage in pure js environment
The way the autocomplete element is configured is to update a hidden input with the defined name for submission as part of a form.

Add an `onChange` prop that is called when the value changes to allow the component to be used outside of a form context in a page which dynamically updates.