Support externally defined species list in species selector
Allows passing in a set of species options for selection rather than being stuck to the predefined PPL or PIL lists. This is so some extra options can be defined for ROPs only.

Also allows injecting checkbox options into the final "Other" section. Specifically for ROPs this is "Other mammals" and "Other carnivores".